Full refactor & logic checks for Type & Eisenhower tracking complete

+++ app/src/main/java/com/team3000/logit/EntryManager.java
@@ -1,7 +1,6 @@
 package com.team3000.logit;
 
 import android.app.Activity;
-import android.content.Intent;
 import android.util.Log;
 import android.widget.Toast;
 
@@ -17,8 +16,8 @@ import com.google.firebase.firestore.FirebaseFirestore;
 import com.google.firebase.firestore.QuerySnapshot;
 import com.google.firebase.firestore.SetOptions;
 
-import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Locale;
 import java.util.Map;
 
@@ -175,33 +174,11 @@ public class EntryManager {
         firestore.document(dbPath).delete().addOnCompleteListener(listener);
     }
 
-    protected void updateTracker(String trackType, String updateDir, String oriDir) {
-        if ("".equals(oriDir)) {
+    protected void updateTracker(String trackType, String updateDir) {
             String trackerPath = String.format("users/%s/%s", user.getUid(), trackType);
             Map<String, String> pathData = new HashMap<>();
             pathData.put("entryPath", updateDir);
             firestore.collection(trackerPath).add(pathData);
-        }
-
-//        String trackerPath = String.format("users/%s", user.getUid());
-//        firestore.document(trackerPath).get().addOnCompleteListener(new OnCompleteListener<DocumentSnapshot>() {
-//            @Override
-//            public void onComplete(@NonNull Task<DocumentSnapshot> task) {
-//                ArrayList<String> temp;
-//                if (task.getResult().get(trackType) == null) {
-//                    temp = new ArrayList<>();
-//                    Log.d(TAG, "New " + trackType + " array added");
-//                } else {
-//                    temp = (ArrayList<String>) task.getResult().get(trackType);
-//                    if (temp.contains(oriDir)) {
-//                        temp.remove(oriDir);
-//                    }
-//                    Log.d(TAG, trackType + " array updated");
-//                }
-//                temp.add(updateDir);
-//                firestore.document(trackerPath).update(trackType, temp);
-//            }
-//        });
     }
 
     protected void deleteFromTracker(String trackType, String entryDir) {
@@ -210,20 +187,12 @@ public class EntryManager {
                 .get().addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
                     @Override
                     public void onComplete(@NonNull Task<QuerySnapshot> task) {
-                        String trackerId = task.getResult().getDocuments().get(0).getId();
-                        firestore.document(trackerPath + "/" + trackerId).delete();
+                        List<DocumentSnapshot> ds = task.getResult().getDocuments();
+                        if (!ds.isEmpty()) {
+                            String trackerId = ds.get(0).getId();
+                            firestore.document(trackerPath + "/" + trackerId).delete();
+                        }
                     }
                 });
-
-//        String trackerPath = String.format("users/%s", user.getUid());
-//        firestore.document(trackerPath).get().addOnCompleteListener(new OnCompleteListener<DocumentSnapshot>() {
-//            @Override
-//            public void onComplete(@NonNull Task<DocumentSnapshot> task) {
-//                ArrayList<String> temp = (ArrayList<String>)task.getResult().get(trackType);
-//                temp.remove(entryDir);
-//                firestore.document(trackerPath).update(trackType, temp);
-//                Log.d(TAG, entryDir + " deleted from array " + trackType);
-//            }
-//        });
     }
 }
